+++ frontend/openchat-agent/src/services/common/chatMappers.ts
@@ -171,6 +171,10 @@ export function messageContent(candid: ApiMessageContent, sender: string): Messa
         // TODO
         throw new Error();
     }
+    if ("Custom" in candid) {
+        // TODO
+        throw new Error();
+    }
     throw new UnsupportedValueError("Unexpected ApiMessageContent type received", candid);
 }